requirements.txt
numpy
openslide-python
Pillow==8.0.1
scipy
scikit-image
h5py